# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/open-api/v1/orders": {
      "post": {
        "description": "商户创建收款订单，返回收银台链接供用户支付。需要 HMAC-SHA256 签名认证。",
        "operationId": "OpenApiController_createOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOpenApiOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "订单创建成功，返回 orderNo 和 cashierUrl"
          },
          "401": {
            "description": "KB401 签名验证失败"
          },
          "403": {
            "description": "KB717 应用已禁用"
          }
        },
        "security": [
          {
            "app-id": []
          }
        ],
        "summary": "创建收款订单",
        "tags": [
          "开放 API"
        ]
      }
    },
    "/open-api/v1/orders/{orderNo}": {
      "get": {
        "description": "根据平台订单号查询订单状态和详情",
        "operationId": "OpenApiController_getOrder",
        "parameters": [
          {
            "name": "orderNo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回订单详情"
          },
          "404": {
            "description": "KB603 订单不存在"
          }
        },
        "security": [
          {
            "app-id": []
          }
        ],
        "summary": "查询订单详情",
        "tags": [
          "开放 API"
        ]
      }
    },
    "/open-api/v1/refunds": {
      "post": {
        "description": "对已支付订单发起全额或部分退款",
        "operationId": "OpenApiController_refund",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefundDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "退款成功"
          },
          "400": {
            "description": "KB713 订单不可退 / KB715 退款金额无效"
          },
          "403": {
            "description": "KB403 无权操作该订单"
          }
        },
        "security": [
          {
            "app-id": []
          }
        ],
        "summary": "申请退款",
        "tags": [
          "开放 API"
        ]
      }
    },
    "/open-api/v1/stats": {
      "get": {
        "summary": "查询商户收单统计",
        "description": "查询当前应用所属商户的收单统计（成功笔数/金额/手续费/退款），金额单位：元。需要 HMAC-SHA256 签名认证。",
        "operationId": "OpenApiController_stats",
        "responses": {
          "200": {
            "description": "返回收单统计（单位：元）"
          }
        },
        "security": [
          {
            "app-id": []
          }
        ],
        "tags": [
          "开放 API"
        ]
      }
    },
    "/qr-codes/fixed": {
      "post": {
        "description": "创建指定金额的静态收款码",
        "operationId": "QrCodesController_createFixedCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFixedCodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "收款码创建成功"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "创建固定金额收款码",
        "tags": [
          "收款码"
        ]
      }
    },
    "/qr-codes/personal": {
      "get": {
        "description": "获取或创建个人动态收款码",
        "operationId": "QrCodesController_getPersonalCode",
        "parameters": [],
        "responses": {
          "200": {
            "description": "返回收款码信息"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "获取个人收款码",
        "tags": [
          "收款码"
        ]
      }
    },
...
    "/users/bind-email": {
      "post": {
        "description": "校验验证码后更新用户邮箱",
        "operationId": "UsersController_bindEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "绑定成功"
          },
          "400": {
            "description": "KB224 验证码错误 / KB223 邮箱已被绑定"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "绑定/换绑邮箱",
        "tags": [
          "用户"
        ]
      }
    },
    "/users/bind-phone": {
      "post": {
        "description": "校验短信验证码后更新用户手机号",
        "operationId": "UsersController_bindPhone",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindPhoneDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "绑定成功"
          },
          "400": {
            "description": "KB224 验证码错误 / KB222 手机号已被绑定"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "绑定/换绑手机号",
        "tags": [
          "用户"
        ]
      }
    },
    "/users/change-password": {
      "post": {
        "description": "验证原登录密码后设置新密码",
        "operationId": "UsersController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "修改成功"
          },
          "401": {
            "description": "KB221 原登录密码错误"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "修改登录密码",
        "tags": [
          "用户"
        ]
      }
    },
    "/users/login-logs": {
      "get": {
        "description": "获取当前用户最近 30 天登录记录",
        "operationId": "UsersController_loginLogs",
        "parameters": [],
        "responses": {
          "200": {
            "description": "返回登录日志列表"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "查询登录日志",
        "tags": [
          "用户"
        ]
      }
    },
    "/users/me": {
      "get": {
        "operationId": "UsersController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "返回用户安全资料（不含密码）"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "获取当前用户信息",
        "tags": [
          "用户"
        ]
      },
      "patch": {
        "description": "更新昵称、头像等基础信息",
        "operationId": "UsersController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "更新当前用户资料",
        "tags": [
          "用户"
        ]
      }
    },
    "/users/reset-pay-password": {
      "post": {
        "description": "验证旧支付密码后设置新密码",
        "operationId": "UsersController_resetPayPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPayPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "重置成功"
          },
          "400": {
            "description": "KB208 支付密码错误"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "重置支付密码",
        "tags": [
          "用户"
        ]
      }
    },
    "/users/verify-identity": {
      "post": {
        "description": "提交身份证姓名+号码进行实名验证",
        "operationId": "UsersController_verifyIdentity",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyIdentityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "提交成功，进入审核"
          },
          "400": {
            "description": "KB202 已实名 / KB203 审核中"
          }
        },
        "security": [
          {
            "user-auth": []
          }
        ],
        "summary": "提交实名认证",
        "tags": [
          "用户"
        ]
      }
    },
    "/webhooks/recharge/{channel}": {
      "post": {
        "description": "支付渠道收单结果回调通知（由渠道服务器调用），语义为 PaymentOrder 支付结果回调。",
        "operationId": "WebhooksController_rechargeCallback",
        "parameters": [
          {
            "name": "channel",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "处理成功"
          }
        },
        "summary": "收单回调",
        "tags": [
          "Webhooks"
        ]
      }
    },
    "/webhooks/refund/{channel}": {
      "post": {
        "description": "退款渠道回调通知",
        "operationId": "WebhooksController_refundCallback",
        "parameters": [
          {
            "name": "channel",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "处理成功"
          }
        },
        "summary": "退款回调",
        "tags": [
          "Webhooks"
        ]
      }
    }
  },
  "info": {
    "title": "科佰支付 KeBaiPay API",
    "description": "合规聚合支付技术服务商接口文档：资金绝不过平台，仅做收单、退款与对账聚合。\n\n## 认证方式\n- **用户接口**：使用 JWT Bearer Token 认证\n- **商户开放 API**：使用 HMAC-SHA256 签名认证\n- **管理后台**：使用管理员 JWT Token 认证\n\n## 错误码规范\n所有错误返回统一格式：`{ \"statusCode\": number, \"message\": \"KBxxx 错误描述\" }`\n\n## 金额说明\n数据库存储单位为「分」，接口传输单位为「元」\n\n> 已下线接口（v0.3.x 精简）：多平台对账聚合（/admin/channel-reconciliation/*）、AI 风控审计（/risk-audit/*、/admin/risk-audit/*）、自定义风控规则（/admin/risk-rules/custom、/risk-rules/custom）。对账以 /admin/reconciliation 为准。",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "user-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "用户/管理员 JWT Token"
      },
      "app-id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-App-Id",
        "description": "商户应用 App ID"
      }
    },
    "schemas": {
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {}
      },
      "VerifyIdentityDto": {
        "type": "object",
        "properties": {}
      },
      "ResetPayPasswordDto": {
        "type": "object",
        "properties": {}
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {}
      },
      "BindPhoneDto": {
        "type": "object",
        "properties": {}
      },
      "BindEmailDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateFixedCodeDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateOpenApiOrderDto": {
        "type": "object",
        "properties": {}
      },
      "RefundDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}